Cleanup fix, mainly removing get-injectable
Added a summaryoutput object for zone and parcel output in order to get rid of get injectable in that part of the code - probably has bugs at this point.

Also adding extensive doc strings and moving a few functions into injectables in datasources.py.

More to come, but I want to move this to the server for testing.