sometimes there are vacant units for buildings that don't make it out of the merge
basically the building DOES exist, but when it's merged to nodes, some of the node_ids are missing because there are parcel_ids in the building table that don't exist in the parcels.index.

anyway, at this point we deal with the issue appropriately and print a warning so this shouldn't come up again.