fix memory error by eliminating copy of jobs table
I was getting a memory error when setting jobs to unplaces when they are in buildings which get redeveloped.  This code has a copy in it - I mean calling the orca method "to_frame" does a copy, so instead of calling that, use the .local dataframe directly which saves a very large copy.  This wasn't a huge performance hit, just seemed to be a bottleneck in terms of memory use.